projects/7/backend_api/serializers.py
from rest_framework import serializers
from . import models


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.ModelBook
        fields = '__all__'


class BookCategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = models.ModelBookCategory
        fields = '__all__'
